Do not assume is_local…() to avoid running collectstatic, but better instead: test if remote environment has DEBUG=True, in which case Django is much more likely to serve static files directly.

+++ sparks/django/fabfile.py
@@ -1124,6 +1124,11 @@ def collectstatic(remote_configuration=None, fast=True):
     """ Run the Django collectstatic management command. If :param:`fast`
         is ``False``, the ``STATIC_ROOT`` will be erased first. """
 
+    if remote_configuration.django_settings.DEBUG:
+        LOGGER.info('NOT running collectstatic on %s because `DEBUG=True`.',
+                    env.host_string)
+        return
+
     if not fast:
         with cd(env.root):
             run('rm -rf "{0}"'.format(
@@ -1297,10 +1302,7 @@ def runable(fast=False, upgrade=False):
     execute_or_not(compilemessages, sparks_roles=('web', 'worker',
                    'worker_low', 'worker_medium', 'worker_high'))
 
-    if not is_local_environment():
-        # In debug mode, Django handles the static contents via a dedicated
-        # view. We don't need to create/refresh/maintain the global static/ dir.
-        execute_or_not(collectstatic, fast=fast, sparks_roles=('web', ))
+    execute_or_not(collectstatic, fast=fast, sparks_roles=('web', ))
 
 
 @task(aliases=('fast', 'fastdeploy', ))
